+++ src/main/java/com/foodservice/service/CustomerService.java
@@ -6,5 +6,5 @@ import java.util.List;
 
 public interface CustomerService {
     List<CustomerDTO> getAllCustomers();
-    CustomerDTO getCustomerById(Long id);
+    CustomerDTO getCustomerById(Integer id);
 }
